// src/collections/filters/types.ts
export type GeoRangeFilter = {
  latitude: number;
  longitude: number;
  distance: number;
};

export type FilterValueType = PrimitiveFilterValueType | PrimitiveListFilterValueType;

export type PrimitiveFilterValueType = number | string | boolean | Date | GeoRangeFilter;
export type PrimitiveListFilterValueType = number[] | string[] | boolean[] | Date[];